requirements.txt
langchain==0.3.15
langchain-core==0.3.31
langgraph==0.2.63
groq==0.15.0
faiss-cpu==1.7.3
streamlit==1.41.1
python-dotenv
requests
sentence-transformers
pytest
coverage
numpy
setuptools